Revert AddComponent Clear (#76)
* Revert "Clear AddComponent function (#75)"

This reverts commit d30d6917cbc35f590c6d18ea521c0058af8f2d5c.

* Fix AddComponent templates

+++ Breakout/Script/BrickController.cpp
@@ -6,7 +6,7 @@ using namespace Learning2DEngine::Render;
 using namespace Learning2DEngine::System;
 using namespace Learning2DEngine::Physics;
 
-BrickController::BrickController(GameObject* gameObject, int x, int y, bool isSolid)
+BrickController::BrickController(GameObject* gameObject, size_t x, size_t y, bool isSolid)
     : Component(gameObject), mapCoorX(x), mapCoorY(y), isSolid(isSolid), collider(nullptr), renderer(nullptr)
 {
